refactor: sign in page enhancements
